refactor(tools,risk-gating): fix 2 domain->application layering violations (#738)
GAP-A-21: two real hexagonal-architecture violations where domain imports
from application:

1. tools/domain/tool_trait.rs imported application::dto::{ToolInput, ToolResult}
   -> moved ToolInput/ToolResult/SideEffect to tools/domain/types.rs (domain
   value objects). application/dto/mod.rs now re-exports them (correct
   application -> domain direction); all 19 downstream users keep compiling.
   Contract check updated to verify the new location.

2. risk_gating/domain/gate_state.rs imported application::dto::PendingGate
   -> PendingGate now lives in risk_gating/domain/gate_state.rs; the
   application DTO module re-exports it for ResolveGateOutput.

The 15 infra->application DTO imports are the correct hexagonal direction
and are untouched (out of scope).

Verification: engine 1998 tests pass; tool-system 20/20 + risk-gating 18/18
contract checks pass; architecture conformance passes; domain dirs contain
no application imports; clippy --all-targets -D warnings clean.
